+++ lib/plugins/deploy.js
@@ -1,8 +1,8 @@
 'use strict';
 
-const generateAnalyticsPayload = require('../../utils/analytics/generatePayload');
-const { report: analyticsReport } = require('../../utils/analytics');
-const processBackendNotificationRequest = require('../../utils/processBackendNotificationRequest');
+const generateAnalyticsPayload = require('../utils/analytics/generatePayload');
+const { report: analyticsReport } = require('../utils/analytics');
+const processBackendNotificationRequest = require('../utils/processBackendNotificationRequest');
 
 class Deploy {
   constructor(serverless, options) {
