+++ lib/plugins/index.js
@@ -6,7 +6,7 @@ module.exports = [
   require('./create/create.js'),
   require('./install/install.js'),
   require('./package/package.js'),
-  require('./deploy/deploy.js'),
+  require('./deploy.js'),
   require('./invoke/invoke.js'),
   require('./info/info.js'),
   require('./logs/logs.js'),
